Initial implementation of disk backed queue.
Thread-safe FIFO queue that uses a muxy filesystem to
write overflow elements to disk. If memory capacity is not exceeded
then all operations occur in memory. {@link #put(Object)}
requests are designed to return quickly and write to disk
asynchronously if one or more background threads have been configured.
All requests of the {@link DiskBackedQueueInternals#get(long, TimeUnit)}
family perform synchronous reads from the disk. A {@link #close()}
will write the contents of the queue to the muxy filesystem.

Nearly all parameter are required to the Builder object.
This is a deliberate design decision to discourage a reliance of magical
default values that lead to behavior surprises for your application.

Single threaded unit tests are passing. Multithreaded test is failing
on IOException for uninitialized muxy stream. I am probably doing something
silly that is easily fixed.

Two changes to core muxy implementation. (1) Added MuxFileDirectory#sync()
method. Written with same concurrency protections as WritableMuxFile#sync()
(which are none) so presumed to be safe. (2) MuxyEventListener is described
as optional in the Javadoc but a few places assumed it was always defined.
Fixed those instances.